+++ src/de/uni_koblenz/jgralab/greql/funlib/collections/Difference.java
@@ -94,7 +94,7 @@ public class Difference extends Function {
 	}
 
 	@Override
-	public long getEstimatedCardinality(int inElements) {
+	public long getEstimatedCardinality(long inElements) {
 		return inElements;
 	}
 }
