+++ src/de/uni_koblenz/jgralab/greql/funlib/collections/EntrySet.java
@@ -68,7 +68,7 @@ public class EntrySet extends Function {
 	}
 
 	@Override
-	public long getEstimatedCardinality(int inElements) {
+	public long getEstimatedCardinality(long inElements) {
 		return inElements;
 	}
 }
