+++ src/de/uni_koblenz/jgralab/greql/funlib/collections/ToSet.java
@@ -70,7 +70,7 @@ public class ToSet extends Function {
 	}
 
 	@Override
-	public long getEstimatedCardinality(int inElements) {
+	public long getEstimatedCardinality(long inElements) {
 		return inElements;
 	}
 }
