+++ src/de/uni_koblenz/jgralab/greql/funlib/collections/Union.java
@@ -98,7 +98,7 @@ public class Union extends Function {
 	}
 
 	@Override
-	public long getEstimatedCardinality(int inElements) {
+	public long getEstimatedCardinality(long inElements) {
 		return inElements;
 	}
 }
